+++ db/migrate/20210128192141_add_external_id_to_users_positions.rb
@@ -1,0 +1,5 @@
+class AddExternalIdToUsersPositions < ActiveRecord::Migration[6.1]
+  def change
+    add_column :users_positions, :external_id, :string
+  end
+end